[seeder] Refactor squad and student seeder for development (#96)
* refactor(seeder): Cleanup and make it easier to development test

Assign only verified student to be leader of squad. If all squad was
have a leader, assign remaining student to be member of random squad

This will make it easier to develop the dashboard, because the squad
will definitely have a leader, and verified students will definitely be
included in the squad.

* refactor(seeder): Remove inviteSquad seeder in call list

Because this is unused for development